Extract matcher constant and use in precondition; apply formatter
