+++ src/test/java/org/openrewrite/rewrite/CorrectlySpacedDescriptionsTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.rewrite.strings;
+package org.openrewrite.rewrite;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
